requirements.txt
boto3==1.26.155
Pillow==9.5.0
rembg==2.0.43